update dynamics plot to support ploting true parameters
