+++ dynamo/plot/__init__.py
@@ -1,8 +1,8 @@
 """Mapping Vector Field of Single Cells
 """
 
-from .preprocess import show_fraction, phase_portraits, feature_genes, variance_explained
-from .dynamics import dynamics
+from .preprocess import show_fraction, feature_genes, variance_explained
+from .dynamics import phase_portraits, dynamics
 from .scVectorField import cell_wise_velocity, grid_velocity, stremline_plot, line_integral_conv # , plot_LIC_gray
 from .scPotential import show_landscape
 from .scatters import scatters
